refactor: refactor router
