+++ packages/iceworks-client/src/pages/Setting/components/SubMenu/index.module.scss
@@ -1,0 +1,4 @@
+.settingSubMenu {
+  display: flex;
+  height: 100%;
+}
